requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-groq
langchain-community
langchain-huggingface
faiss-cpu
pymupdf
sentence-transformers
python-dotenv
pydantic
python-docx
